Safari-611.1.10/LayoutTests/editing/pasteboard/do-not-reveal-selection-after-programmatic-scroll-expected.txt

This test verifies that we do not scroll to reveal the selection after pasting images in an editable area, if the page programmatically scrolls the editor. To run the test manually, copy the selected images and paste into the red editable container above. The container's scroll position should remain at 0.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS allImagesAreDoneLoading() became true
PASS editor.scrollTop is 0
PASS successfullyParsed is true

TEST COMPLETE
Here's to the crazy ones, the misfits, the rebels, the trouble makers, the round pegs in the square holes, the ones who see things differently. There not fond of rules, and they have no respect for the status quo, you can quote then, disagree with them, glorify or vilify them, about the only thing you can't do is ignore them. Because they change things. They push the human race forward. And while some may see them as the crazy ones, we see genius. Because the people who are crazy enough to think they can change the world are the ones who do.
